requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.8.2
pandas==2.2.2
openpyxl==3.1.5
xlrd==2.0.1
python-docx==1.1.2
mt-940==4.30.0
opendataloader-pdf>=1.11.2
pypdf==4.2.0
PyMuPDF==1.24.9
pytesseract==0.3.10
Pillow==10.4.0
numpy==2.0.1
rapidocr-onnxruntime==1.4.4
